More code reorganization
Moved the PolicyEvaluator and other structs/functions into a dedicated
workspace.

This allows to resulse the PolicyEvaluator code between policy-server
and policy-testdrive.
