+++ .github/workflows/tests.yml
@@ -30,6 +30,7 @@ jobs:
       - uses: actions-rs/cargo@v1
         with:
           command: test
+          args: --workspace
 
   fmt:
     name: Rustfmt
