+++ crates/policy-evaluator/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "policy-evaluator"
+version = "0.1.0"
+authors = ["Flavio Castelli <fcastelli@suse.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+chimera-kube-policy-sdk = { git = "https://github.com/chimera-kube/policy-sdk-rust.git", branch = "main" }
+hyper = { version = "0.14" }
+json-patch = "0.2.6"
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = "0.8.15"
+wapc = "0.10.1"
+wasmtime-provider = "0.0.2"
