+++ crates/policy-evaluator/src/policy.rs
@@ -7,7 +7,7 @@ use std::collections::HashMap;
 use std::fs::File;
 
 #[derive(Deserialize, Debug, Clone)]
-pub(crate) struct Policy {
+pub struct Policy {
     pub url: String,
 
     #[serde(skip)]
@@ -18,7 +18,7 @@ pub(crate) struct Policy {
 }
 
 impl Policy {
-    pub(crate) fn settings(&self) -> Mapping {
+    pub fn settings(&self) -> Mapping {
         self.extra_fields
             .get("settings")
             .map(|v| {
@@ -38,7 +38,7 @@ impl Policy {
 // and Policy as values. The key is the name of the policy as provided by the user
 // inside of the configuration file. This name is used to build the API path
 // exposing the policy.
-pub(crate) fn read_policies_file(path: &str) -> Result<HashMap<String, Policy>> {
+pub fn read_policies_file(path: &str) -> Result<HashMap<String, Policy>> {
     let settings_file = File::open(path)?;
     let ps: HashMap<String, Policy> = serde_yaml::from_reader(&settings_file)?;
     Ok(ps)
