+++ src/main.rs
@@ -14,13 +14,10 @@ mod admission_review;
 mod api;
 mod registry;
 mod server;
-mod utils;
-mod validation_response;
 mod wasm_fetcher;
 mod worker;
 
-mod policies;
-use policies::read_policies_file;
+use policy_evaluator::policy::read_policies_file;
 
 mod wasm;
 use crate::wasm::EvalRequest;
